api/requirements.txt
databases[asyncpg]
fastapi==0.112.0
numpy
orjson
pandas
python-dotenv
python-multipart
uvicorn[standard]
